Add Direct Reference addon project & chip wrapper
In order to make it so that other projects don't have
to work with a publicized version of the dll, I have
started adding wrappers to expose the private fields
and classes that should be most pertinent to adding
opcodes.

I have replaced the arg count field with an accept
method on the off chance someone wants a variable arg
command. The 'error' example comes to mind as one such
since you might want to optionally set the Setting field
with an error code. This would allow for that as well as
a no-arg version that does not change the Setting field.

I do feel like that should be used sparingly however, as it
seems out of the spirit.